moved test under the correct folder
